requirements.txt
# ── Core PDF extraction ──────────────────────────────────────────────────────
# Primary extractor: fast, reliable for digital PDFs
PyMuPDF>=1.23.0

# Fallback extractor: better for tables and multi-column layouts
pdfplumber>=0.10.0

# ── OCR (optional) ───────────────────────────────────────────────────────────
# Required ONLY if you process scanned/image PDFs
# Also needs Tesseract OCR binary installed separately:
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   Linux:   sudo apt-get install tesseract-ocr
#   macOS:   brew install tesseract
pytesseract>=0.3.10
Pillow>=10.0.0

# ── CLI & console output ─────────────────────────────────────────────────────
click>=8.1.0
rich>=13.0.0

# ── Configuration ─────────────────────────────────────────────────────────────
PyYAML>=6.0

# ── Semantic search (ChromaDB + embeddings) ───────────────────────────────────
# Required for semantic search and MCP server functionality
chromadb>=0.5.0
sentence-transformers>=2.2.0

# ── MCP server (Claude Code integration) ─────────────────────────────────────
mcp>=1.0.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
